// src/app/core/model/types-reasignacion.ts
export enum CodigoCatalogo {
    DescripcionMaestroProceso = 121,
    EstadoDesarrollo = 112
}

export enum CodigoEstadoDesarrollo {
    Pendiente = 1,
    Iniciado = 2,
    Finalizado = 3,
    Cancelado = 4
}

export enum CodigoEstadoCalificacion {
    Pendiente = 1,
    Apto = 2,
    NoApto = 3,
    Observado = 4
}

export enum CodigoEstadoAdjudicacion {
    Pendiente = 1,
    Adjudicado = 2,
    NoAdjudicado = 3,
    Observado = 4
}

export enum codigoEstadoValidacionPlaza {
    Pendiente = 1,
    Validado = 2,
    PrePublicado = 3,
    Rechazado = 4,
    Aprobado = 5,
    Publicado = 6,
    Aperturado = 7,
    PrePublicandoPlazas = 8,
    PublicandoPlazas = 9
}

export enum codigoResultadoFinal {
    Pendiente = 1,
    Adjudicado = 2,
    Desierto = 3
}

export enum CodigoDescripcionMaestroProceso {
    PrePublicacionPlazasReasignacion = 27,
    PorInteresPersonalPorUnidadFamiliar = 25
}

export enum TablaRegimenLaboral {
    Ley29944LRM = 1
}